refactor(npegl): Go back to subclassing StatsProxy
The NPEGL stats row merging code could not be loaded properly due to global variables and order of imports. This was fixed in d59a9b5 by inventing a new `stats_rewrite` config option. Now, StatsProxy is wrapped in a Factory helper, so the import order is less tangled. So this commit goes back to the old way.